pyproject.toml
[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[project]
name = "az-table-catalog"
version = "1.0.0"
authors = [
  { name="C. Shaun Wagner", email="cs@kainaw.com" },
]
description = "A resilient, event-sourced Azure Table Storage catalog index"
readme = "README.md"
license = { file = "LICENSE" }
requires-python = ">=3.10"
dependencies = [
    "azure-data-tables>=12.0.0",
]
classifiers = [
    "Programming Language :: Python :: 3",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Development Status :: 4 - Beta",
    "Intended Audience :: Developers",
]

[project.urls]
"Homepage" = "https://github.com/kainaw/az-table-catalog"
"Bug Tracker" = "https://github.com/kainaw/az-table-catalog/issues"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
include = ["az_table_catalog*"]